move extractOverrides logic to separate file
